+++ HISTORY.rst
@@ -13,6 +13,7 @@ Unreleased
 * ALCYN-1494: Fix a regression introduced by ALCYN-465 that was generating errors in imports of sales and purchases
 * ALCYN-1485: SO import - Group shipping backorders by partners
 * ALCYN-126: On import switch discounts when both are set, or when product has GMA price category
+* ALCYN-1500: Create a migration location for a fake stock to use as a source on imported Sale orders
 
 **Features and Improvements**